Create LICENSE
Written permission from the Roman Catholic Archdiocese of Vancouver:

"The Roman Catholic Archdiocese of Vancouver (RCAV) releases for Open-Source development the “Refugees Support” application, developed in 2020-2021 on the Salesforce platform by its contracted employees Matthew Martin and Jack Ong.  The RCAV understands that, upon releasing the Refugees Support application to the open-source community, it and any other entity will be able to use the application and contribute to the application’s ongoing development but will not maintain proprietary ownership of the application nor will it be compensated financially for initial or ongoing work in any way." Signed: Sean O'Brien, Director of Finance, RCAV, on Thursday, Dev 9, 2021.